Minor improvements
- Remove unneccessary global function
- Get rid of mock dependency
